Refactor code to resolve circular dependency problem
