+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/controller/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.avdo.spring.app;
+package com.avdo.spring.app.controller;
 
 import lombok.Getter;
 import lombok.Setter;
