+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/controller/UserController.java
@@ -1,6 +1,5 @@
 package com.avdo.spring.app.controller;
 
-import com.avdo.spring.app.LoginResponse;
 import com.avdo.spring.app.dto.CreateUserRequest;
 import com.avdo.spring.app.dto.LoginUserRequest;
 import com.avdo.spring.app.entity.User;
